requirements.txt
Keras==2.2.4
scikit_learn==0.20.3
numpy==1.16.2
opencv-python==4.5.5.62
scikit-image==0.19.2
Flask==2.0.3
Pillow==9.0.1
flask-ngrok==0.0.25